make pcguard work
